Struct jnet::coap::Message

pub struct Message<BUFFER> where
    BUFFER: AsRef<[u8]>,  { /* fields omitted */ }

CoAP (version 1) message

Methods

impl<B> Message<B> where
B: AsRef<[u8]>,

pub fn parse(bytes: B) -> Result<Self, B>

Parses bytes into a CoAP message

pub fn get_version(&self) -> u8

Returns the Version field of the header

As per RFC 7252 this always returns 1

pub fn get_type(&self) -> Type

Returns the Type field of the header

pub fn get_token_length(&self) -> u8

Returns the Token Length (TKL) field of the header

As per RFC 7252 this always returns a value in the range 0..=8

pub fn get_code(&self) -> Code

Returns the Code field of the header

pub fn get_message_id(&self) -> u16

Returns the Message ID field of the header

pub fn token(&self) -> &[u8]

View into the Token field of the header

Important traits for Options<'a>
impl<'a> Iterator for Options<'a> type Item = Option<'a>;

pub fn options(&self) -> Options

Returns an iterator over the options of this message

pub fn payload(&self) -> &[u8]

View into the payload

pub fn len(&self) -> u16

Returns the length (header + data) of the CoAP message

pub fn as_bytes(&self) -> &[u8]

Returns the byte representation of this message

impl<B> Message<B> where
B: AsRef<[u8]> + AsMut<[u8]>,

pub fn new(buffer: B, token_length: u8) -> Self

Transforms the given buffer into a CoAP message

This constructor sets the following header fields

Version = 1

Token Length = token_length

NOTE The CoAP message will span the whole buffer.

Panics

This constructor panics if

token_length is NOT in the range 0..=8.

The buffer is not large enough to contain the CoAP header

pub fn add_option(&mut self, number: OptionNumber, value: &[u8])

Adds an option to this message

HEADS UP This method will cause the first bytes of the payload to be lost

Panics

This method panics

if number is smaller than the highest option number already contained in the message

if there's no space in the message to add the option

pub fn clear_options(&mut self)

Removes all the options this message has

pub fn token_mut(&mut self) -> &mut [u8]

Mutable view into the Token field

pub fn payload_mut(&mut self) -> &mut [u8]

Mutable view into the payload

pub fn set_code<C>(&mut self, code: C) where
C: Into<Code>,

Sets the Code field of the header

pub fn set_message_id(&mut self, id: u16)

Sets the Message ID field of the header

pub fn set_type(&mut self, ty: Type)

Sets the Type field of the header

impl<B> Message<B> where
B: AsRef<[u8]> + Resize,

pub fn truncate(&mut self, len: u16)

Truncates the payload to the specified length

impl<B> Message<B> where
B: AsRef<[u8]> + AsMut<[u8]> + Resize,

pub fn set_payload(&mut self, data: &[u8])

Fills the payload with the given data and adjusts the length of the CoAP message