Struct jnet::ether::Frame

pub struct Frame<BUFFER> where
    BUFFER: AsSlice<Element = u8>,  { /* fields omitted */ }

Layer 2 Ethernet frame

Structure

MAC destination. 6 bytes

MAC source. 6 bytes

Ethertype. 2 bytes

Payload. 46-1500 bytes (*)

Frame check sequence. 4 bytes (*)

(*) This frame representation does NOT include the frame check sequence nor (zero) pads the payload to the minimum size of 46 bytes.

Methods

impl<B> Frame<B> where
B: AsSlice<Element = u8>,

pub fn new(buffer: B) -> Self

Creates a new Ethernet frame from the given buffer

pub fn parse(bytes: B) -> Result<Self, B>

Parses bytes into an Ethernet frame

pub fn get_destination(&self) -> Addr

Returns the Destination field of the header

pub fn get_source(&self) -> Addr

Returns the Source field of the header

pub fn get_type(&self) -> Type

Returns the Type field of the header

pub fn payload(&self) -> &[u8]

View into the payload

pub fn as_bytes(&self) -> &[u8]

Returns the byte representation of this frame

pub fn free(self) -> B

Frees the underlying buffer

pub fn len(&self) -> u16

Returns the length (header + data) of this frame

impl<B> Frame<B> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,

pub fn set_destination(&mut self, addr: Addr)

Sets the destination field of the header

pub fn set_source(&mut self, addr: Addr)

Sets the source field of the header

pub fn set_type(&mut self, type_: Type)

Sets the type field of the header

pub fn payload_mut(&mut self) -> &mut [u8]

Mutable view into the payload

impl<B> Frame<B> where
B: AsSlice<Element = u8> + IntoSliceFrom<u8>,

pub fn into_payload(self) -> B::SliceFrom

Returns the payload of this frame

impl<B> Frame<B> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8> + Truncate<u8>,

pub fn arp<F>(&mut self, f: F) where
F: FnOnce(&mut Packet<&mut [u8]>),

Fills the payload with an ARP packet

This method sets the Type field of this frame to ARP, and truncates the length of the frame to fit the ARP packet.

The ARP packet will have its SHA set to the Ethernet frame Source address

impl<B> Frame<B> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8> + Truncate<u16>,

pub fn ipv4<F>(&mut self, f: F) where
F: FnOnce(&mut Packet<&mut [u8], Invalid>),

Fills the payload with an IPv4 packet

This method sets the Type field of this frame to IPv4, recomputes and updates the header checksum of the IPv4 payload, and truncates the length of the frame to fit the IPv4 packet.

pub fn ipv6<F>(&mut self, f: F) where
F: FnOnce(&mut Packet<&mut [u8]>),

Fills the payload with an IPv6 packet