Struct jnet::icmp::Message

pub struct Message<BUFFER, TYPE, CHECKSUM> where
    BUFFER: AsSlice<Element = u8>,
    TYPE: 'static,  { /* fields omitted */ }

ICMP Message

Methods

impl<B> Message<B, EchoRequest, Invalid> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,

pub fn new(buffer: B) -> Self

Transforms the input buffer into a Echo Request ICMP packet

impl<B, E, C> Message<B, E, C> where
B: AsSlice<Element = u8>,
E: Echo,

pub fn get_identifier(&self) -> u16

Returns the Identifier field of the header

pub fn get_sequence_number(&self) -> u16

Returns the Identifier field of the header

impl<B, E> Message<B, E, Invalid> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,
E: Echo,

pub fn set_identifier(&mut self, ident: u16)

Returns the Identifier field of the header

pub fn set_sequence_number(&mut self, seq_no: u16)

Returns the Identifier field of the header

impl<B> Message<B, Unknown, Valid> where
B: AsSlice<Element = u8>,

pub fn parse(bytes: B) -> Result<Self, B>

Parses the input bytes into a

impl<B> Message<B, Unknown, Invalid> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,

pub fn set_type(&mut self, type_: Type)

Sets the Type field of the header

pub fn set_code(&mut self, code: u8)

Sets the Code field of the header

impl<B> Message<B, Unknown, Valid> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,

pub fn set_type(self, type_: Type) -> Message<B, Unknown, Invalid>

Sets the Type field of the header

pub fn set_code(self, code: u8) -> Message<B, Unknown, Invalid>

Sets the Code field of the header

impl<B, C> Message<B, Unknown, C> where
B: AsSlice<Element = u8>,

pub fn downcast<TYPE>(self) -> Result<Message<B, TYPE, C>, Self> where
Self: TryInto<Message<B, TYPE, C>, Error = Self>,

Downcasts this packet with unknown type into a specific type

impl<B, T, C> Message<B, T, C> where
B: AsSlice<Element = u8>,

pub fn get_type(&self) -> Type

Returns the Type field of the header

pub fn get_code(&self) -> u8

Returns the Type field of the header

pub fn payload(&self) -> &[u8]

View into the payload

pub fn len(&self) -> u16

Returns the length (header + data) of this packet

pub fn as_bytes(&self) -> &[u8]

Returns the byte representation of this packet

impl<B, T> Message<B, T, Invalid> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,

pub fn payload_mut(&mut self) -> &mut [u8]

Mutable view into the payload

pub fn update_checksum(self) -> Message<B, T, Valid>

Updates the Checksum field of the header