Struct jnet::icmp::Packet

pub struct Packet<BUFFER, TYPE, CHECKSUM> where
    BUFFER: AsRef<[u8]>,
    TYPE: 'static,  { /* fields omitted */ }

ICMP packet

Methods

impl<B> Packet<B, EchoRequest, Invalid> where
B: AsRef<[u8]> + AsMut<[u8]> + Resize,

pub fn new(buffer: B) -> Self

Transforms the input buffer into a Echo Request ICMP packet

impl<B, E, C> Packet<B, E, C> where
B: AsRef<[u8]>,
E: Echo,

pub fn get_identifier(&self) -> u16

Returns the Identifier field of the header

pub fn get_sequence_number(&self) -> u16

Returns the Identifier field of the header

impl<B, E> Packet<B, E, Invalid> where
B: AsRef<[u8]> + AsMut<[u8]>,
E: Echo,

pub fn set_identifier(&mut self, ident: u16)

Returns the Identifier field of the header

pub fn set_sequence_number(&mut self, seq_no: u16)

Returns the Identifier field of the header

impl<B> Packet<B, Unknown, Valid> where
B: AsRef<[u8]> + Resize,

pub fn parse(bytes: B) -> Result<Self, B>

Parses the input bytes into a

impl<B> Packet<B, Unknown, Invalid> where
B: AsRef<[u8]> + AsMut<[u8]>,

pub fn set_type(&mut self, type_: Type)

Sets the Type field of the header

pub fn set_code(&mut self, code: u8)

Sets the Code field of the header

impl<B> Packet<B, Unknown, Valid> where
B: AsRef<[u8]> + AsMut<[u8]>,

pub fn set_type(self, type_: Type) -> Packet<B, Unknown, Invalid>

Sets the Type field of the header

pub fn set_code(self, code: u8) -> Packet<B, Unknown, Invalid>

Sets the Code field of the header

impl<B, C> Packet<B, Unknown, C> where
B: AsRef<[u8]>,

pub fn downcast<TYPE>(self) -> Result<Packet<B, TYPE, C>, Self> where
Self: TryInto<Packet<B, TYPE, C>, Error = Self>,

Downcasts this packet with unknown type into a specific type

impl<B, T, C> Packet<B, T, C> where
B: AsRef<[u8]>,

pub fn get_type(&self) -> Type

Returns the Type field of the header

pub fn get_code(&self) -> u8

Returns the Type field of the header

pub fn payload(&self) -> &[u8]

View into the payload

pub fn len(&self) -> u16

Returns the length (header + data) of this packet

pub fn as_bytes(&self) -> &[u8]

Returns the byte representation of this packet

impl<B, T> Packet<B, T, Invalid> where
B: AsRef<[u8]> + AsMut<[u8]>,

pub fn payload_mut(&mut self) -> &mut [u8]

Mutable view into the payload

pub fn update_checksum(self) -> Packet<B, T, Valid>

Updates the Checksum field of the header