Struct jnet::icmpv6::Message

pub struct Message<BUFFER, TYPE> where
    BUFFER: AsSlice<Element = u8>,  { /* fields omitted */ }

ICMPv6 Message

Methods

impl<B, T> Message<B, T> where
B: AsSlice<Element = u8>,

pub fn get_type(&self) -> Type

Reads the 'Type' field

pub fn get_code(&self) -> u8

Reads the 'Code' field

pub fn get_checksum(&self) -> u16

Reads the 'Checksum' field

pub fn as_bytes(&self) -> &[u8]

Returns the byte representation of this frame

pub fn verify_checksum(&self, src: Addr, dest: Addr) -> bool

Verifies the 'Checksum' field

pub fn free(self) -> B

Returns the underlying buffer

impl<B, T> Message<B, T> where
B: AsMutSlice<Element = u8>,

pub fn update_checksum(&mut self, src: Addr, dest: Addr)

Recomputes and updates the 'Checksum' field

impl<B> Message<B, Unknown> where
B: AsSlice<Element = u8>,

pub fn parse(bytes: B) -> Result<Self, B>

Parses the bytes as an ICMPv6 message NOTE this function does not validate the message checksum

impl<B> Message<B, Unknown> where
B: AsSlice<Element = u8>,

pub fn downcast<TYPE>(self) -> Result<Message<B, TYPE>, Message<B, Unknown>> where
Self: TryInto<Message<B, TYPE>, Error = Self>,

Downcasts this packet with unknown type into a specific type

impl<B> Message<B, NeighborSolicitation> where
B: AsSlice<Element = u8>,

pub fn get_target(&self) -> Addr

Reads the 'Target Address' field

pub fn get_source_ll(&self) -> Option<&[u8]>

Reads the 'Source Link-layer address' option

impl<B> Message<B, NeighborAdvertisement> where
B: AsMutSlice<Element = u8> + Truncate<u8>,

pub fn neighbor_advertisement(buffer: B, target_ll_opt_size: u8) -> Self

Transforms the input buffer into a Neighbor Advertisement ICMPv6 message

target_ll_opt_size is the size of the 'Target Link-layer Address' option in units of 8 octets. A value of 0 means that the option will be omitted.

All these fields need to be filled by the caller

Override bit

Solicited bit

Router bit

Target Address field

Target Link-layer Address option

impl<B> Message<B, NeighborAdvertisement> where
B: AsSlice<Element = u8>,

pub fn get_router(&self) -> bool

Reads the 'Router' flag

pub fn get_solicited(&self) -> bool

Reads the 'Solicited' flag

pub fn get_override(&self) -> bool

Reads the 'Override' flag

pub fn get_target(&self) -> Addr

Reads the 'Target Address' field

pub fn get_target_ll(&self) -> Option<&[u8]>

Reads the 'Target Link-layer Address' option

impl<B> Message<B, NeighborAdvertisement> where
B: AsMutSlice<Element = u8>,

pub fn set_router(&mut self, router: bool)

Sets the 'Router' flag

pub fn set_solicited(&mut self, solicited: bool)

Sets the 'Solicited' flag

pub fn set_override(&mut self, override_: bool)

Sets the 'Override' flag

pub fn set_target(&mut self, addr: Addr)

Sets the 'Target Address' field

pub fn target_ll_mut(&mut self) -> Option<&mut [u8]>

Mutable view into the 'Target Link-layer address' option

impl<B, E> Message<B, E> where
B: AsSlice<Element = u8>,
E: Echo,

pub fn get_identifier(&self) -> u16

Reads the 'Identifier' field

pub fn get_sequence_number(&self) -> u16

Reads the 'Sequence number' field

pub fn payload(&self) -> &[u8]

Immutable view into the payload of this message

impl<B> Message<B, EchoReply> where
B: AsMutSlice<Element = u8>,

pub fn echo_reply(buffer: B) -> Self

Transforms the input buffer into a Echo Reply ICMPv6 message

pub fn set_identifier(&mut self, id: u16)

Sets the 'Identifier' field

pub fn set_sequence_number(&mut self, seq: u16)

Sets the 'Sequence number' field

impl<B> Message<B, EchoReply> where
B: AsMutSlice<Element = u8> + Truncate<u8>,

pub fn set_payload(&mut self, data: &[u8])

Fills the payload with the given data and adjusts the length of the UDP packet