Struct jnet::ieee802154::Frame

pub struct Frame<BUFFER> where
    BUFFER: AsSlice<Element = u8>,  { /* fields omitted */ }

IEEE 802.15.4 MAC frame

Methods

impl<B> Frame<B> where
B: AsSlice<Element = u8>,

pub fn parse(bytes: B) -> Result<Self, B>

Parses bytes into an 802.15.4 frame

pub fn get_type(&self) -> Type

Reads the 'Frame type' field

pub fn get_security_enabled(&self) -> bool

Reads the 'Security enabled' field

pub fn get_frame_pending(&self) -> bool

Reads the 'Frame pending' field

pub fn get_ack_request(&self) -> bool

Reads the 'Ack. request' field

pub fn get_intra_pan(&self) -> bool

Reads the 'Intra-PAN' field

pub fn get_dest_addr_mode(&self) -> AddrMode

Reads the 'Dest. addressing mode' field

pub fn get_src_addr_mode(&self) -> AddrMode

Reads the 'Source addressing mode' field

pub fn get_sequence_number(&self) -> u8

Reads the 'Sequence number' field

pub fn get_dest_pan_id(&self) -> Option<PanId>

Reads the 'Destination PAN identifier' field

pub fn get_dest_addr(&self) -> Option<Addr>

Reads the 'Destination address' field

pub fn get_src_pan_id(&self) -> Option<PanId>

Reads the 'Source PAN identifier' field

pub fn get_src_addr(&self) -> Option<Addr>

Reads the 'Source address' field

pub fn header(&self) -> &[u8]

Returns an immutable view into the header

pub fn payload(&self) -> &[u8]

Returns an immutable view into the payload

pub fn as_bytes(&self) -> &[u8]

Returns the byte representation of this frame

pub fn free(self) -> B

Frees the underlying buffer

impl<B> Frame<B> where
B: AsMutSlice<Element = u8>,

pub fn data(buffer: B, src_dest: SrcDest) -> Self

Creates a new data frame from the given buffer

pub fn set_ack_request(&mut self, ack: bool)

Sets the 'Ack. request' field to ack

pub fn set_sequence_number(&mut self, seq: u8)

Sets the 'Sequence number' field to seq

pub fn payload_mut(&mut self) -> &mut [u8]

Returns a mutable view into the payload

impl<B> Frame<B> where
B: AsMutSlice<Element = u8> + Truncate<u8>,

pub fn set_payload(&mut self, payload: &[u8])

Fills the payload with the given data and adjusts the length of the frame

pub fn echo_reply<F>(&mut self, src: Addr, dest: Addr, f: F) where
F: FnOnce(&mut Message<&mut [u8], EchoReply>),

Fills the buffer with an 'Echo Reply' ICMPv6 message

pub fn neighbor_advertisement<F>(
&mut self,
src: Addr,
dest: Addr,
target_ll_addr: Option<ExtendedAddr>,
f: F
) where
F: FnOnce(&mut Message<&mut [u8], NeighborAdvertisement>),

Fills the payload with a 'Neighbor Advertisement' ICMPv6 message

pub fn udp<F>(
&mut self,
src_addr: Addr,
src_port: u16,
dest_addr: Addr,
dest_port: u16,
elide_checksum: bool,
f: F
) where
F: FnOnce(&mut UdpPacket<&mut [u8]>),

Fills the payload with a UDP packet