Struct jnet::ipv4::Packet

pub struct Packet<BUFFER, CHECKSUM> where
    BUFFER: AsRef<[u8]>,  { /* fields omitted */ }

IPv4 packet

Methods

impl<B> Packet<B, Valid> where
B: AsRef<[u8]> + Resize,

pub fn parse(bytes: B) -> Result<Self, B>

Parses bytes into an IPv4 packet

impl<B, C> Packet<B, C> where
B: AsRef<[u8]>,

pub fn get_version(&self) -> u8

Returns the version field of the header

pub fn get_ihl(&self) -> u8

Returns the IHL (Internet Header Length) field of the header

pub fn get_dscp(&self) -> u8

Returns the DSCP (Differentiated Services Code Point) field of the header

pub fn get_ecn(&self) -> u8

Returns the ECN (Explicit Congestion Notification) field of the header

pub fn get_total_length(&self) -> u16

Returns the total length field of the header

pub fn len(&self) -> u16

Returns the length (header + data) of this packet

This returns the same value as the get_total_length method

pub fn get_identification(&self) -> u16

Returns the identification field of the header

pub fn get_df(&self) -> bool

Returns the DF (Don't Fragment) field of the header

pub fn get_mf(&self) -> bool

Returns the MF (More Fragments) field of the header

pub fn get_fragment_offset(&self) -> u16

Returns the Fragment Offset field of the header

pub fn get_ttl(&self) -> u8

Returns the TTL (Time To Live) field of the header

pub fn get_protocol(&self) -> Protocol

Returns the protocol field of the header

pub fn get_source(&self) -> Addr

Returns the Source (IP address) field of the header

pub fn get_destination(&self) -> Addr

Returns the Destination (IP address) field of the header

pub fn payload(&self) -> &[u8]

View into the payload

impl<B, C> Packet<B, C> where
B: AsRef<[u8]> + AsMut<[u8]>,

pub fn payload_mut(&mut self) -> &mut [u8]

View into the payload

impl<B, C> Packet<B, C> where
B: AsRef<[u8]> + Resize,

pub fn into_payload(self) -> B

Returns the payload of this frame

impl<B> Packet<B, Invalid> where
B: AsRef<[u8]> + AsMut<[u8]> + Resize,

pub fn new(buffer: B) -> Self

Transforms the given buffer into an IPv4 packet

Most of the header will be filled with sensible defaults:

Version = 4

IHL = 5

DSCP = 0

ECN = 0

Total Length = cmp::min(buffer.len(), u16::MAX)

Identification = 0

DF = true

MF = false

Fragment Offset = 0

TTL = 64

The fields that are left unpopulated are:

Protocol

Checksum

Source

Destination

Panics

This constructor panics if the given buffer is smaller than MIN_HEADER_SIZE

pub fn echo_request<F>(&mut self, f: F) where
F: FnOnce(&mut Packet<&mut [u8], EchoRequest, Invalid>),

Fills the payload with an Echo Request ICMP packet

pub fn udp<F>(&mut self, f: F) where
F: FnOnce(&mut Packet<&mut [u8]>),

Fills the payload with an UDP packet

pub fn truncate(&mut self, len: u16)

Truncates the payload to the specified length

impl<B> Packet<B, Valid> where
B: AsRef<[u8]> + AsMut<[u8]> + Resize,

pub fn truncate(self, len: u16) -> Packet<B, Invalid>

Truncates the payload to the specified length

impl<B> Packet<B, Invalid> where
B: AsRef<[u8]> + AsMut<[u8]>,

pub fn set_version(&mut self, version: u8)

Sets the version field of the header

pub fn set_dscp(&mut self, dscp: u8)

Sets the DSCP (Differentiated Services Code Point) field of the header

pub fn set_ecn(&mut self, ecn: u8)

Sets the ECN (Explicit Congestion Notification) field of the header

pub fn set_identification(&mut self, id: u16)

Sets the identification field of the header

pub fn set_df(&mut self, df: bool)

Sets the DF (Don't Fragment) field of the header

pub fn set_mf(&mut self, mf: bool)

Sets the MF (More Fragments) field of the header

pub fn set_fragment_offset(&mut self, fo: u16)

Sets the Fragment Offset field of the header

pub fn set_ttl(&mut self, ttl: u8)

Sets the TTL (Time To Live) field of the header

pub fn set_protocol(&mut self, proto: Protocol)

Sets the Protocol field of the header

pub fn set_source(&mut self, addr: Addr)

Sets the Source (IP address) field of the header

pub fn set_destination(&mut self, addr: Addr)

Sets the Destination (IP address) field of the header

pub fn update_checksum(self) -> Packet<B, Valid>

Updates the Checksum field of the header

impl<B> Packet<B, Valid> where
B: AsRef<[u8]> + AsMut<[u8]>,

pub fn set_version(self, version: u8) -> Packet<B, Invalid>

Sets the version field of the header

pub fn set_dscp(self, dscp: u8) -> Packet<B, Invalid>

Sets the DSCP (Differentiated Services Code Point) field of the header

pub fn set_ecn(self, ecn: u8) -> Packet<B, Invalid>

Sets the ECN (Explicit Congestion Notification) field of the header

pub fn set_identification(self, id: u16) -> Packet<B, Invalid>

Sets the identification field of the header

pub fn set_df(self, df: bool) -> Packet<B, Invalid>

Sets the DF (Don't Fragment) field of the header

pub fn set_mf(self, mf: bool) -> Packet<B, Invalid>

Sets the MF (More Fragments) field of the header

pub fn set_fragment_offset(self, fo: u16) -> Packet<B, Invalid>

Sets the Fragment Offset field of the header

pub fn set_ttl(self, ttl: u8) -> Packet<B, Invalid>

Sets the TTL (Time To Live) field of the header

pub fn set_protocol(self, proto: Protocol) -> Packet<B, Invalid>

Sets the Protocol field of the header

pub fn set_source(self, addr: Addr) -> Packet<B, Invalid>

Sets the Source (IP address) field of the header

pub fn set_destination(self, addr: Addr) -> Packet<B, Invalid>

Sets the Destination (IP address) field of the header