Struct jnet::ipv6::Packet

pub struct Packet<BUFFER> where
    BUFFER: AsSlice<Element = u8>,  { /* fields omitted */ }

IPv6 packet

Methods

impl<B> Packet<B> where
B: AsSlice<Element = u8>,

pub fn parse(bytes: B) -> Result<Self, ()>

Parses bytes into an IPv6 packet

pub fn get_version(&self) -> u8

Reads the 'Version' field

This always returns 6

pub fn get_traffic_class(&self) -> u8

Reads the 'Traffic Class' field

pub fn get_flow_label(&self) -> u32

Reads the 'Flow Label' field (20 bits)

pub fn get_length(&self) -> u16

Reads the 'Payload length' field

pub fn get_next_header(&self) -> NextHeader

Reads the 'Next Header' field

pub fn get_hop_limit(&self) -> u8

Reads the 'Hop Limit' field

pub fn get_source(&self) -> Addr

Reads the 'Source Address' field

pub fn get_destination(&self) -> Addr

Reads the 'Destination Address' field

pub fn payload(&self) -> &[u8]

Immutable view into the payload

pub fn as_bytes(&self) -> &[u8]

Returns the byte representation of this packet

impl<B> Packet<B> where
B: AsMutSlice<Element = u8>,

pub fn new(buffer: B) -> Self

Transforms the given buffer into an IPv4 packet

Most of the header will be filled with sensible defaults:

Version = 6

Traffic class = 0

Flow label = 0

Length = buffer.len() - HEADER_SIZE

Hop limit = 255

The fields that are left unpopulated are:

Next header

Source address

Destination address

Panics

This constructor panics if

the given buffer is smaller than HEADER_SIZE

the packet would result in a payload length larger than u16::MAX.

pub fn set_traffic_class(&mut self, tc: u8)

Sets the 'Traffic class' field

pub fn set_flow_label(&mut self, fl: u32)

Sets the 'Flow label' field

pub fn set_next_header(&mut self, nh: NextHeader)

Sets the 'Next Header' field

Panics

This function panics if nh is an extension header (currently not supported)

pub fn set_hop_limit(&mut self, hl: u8)

Sets the 'Hop limit' field

pub fn set_source(&mut self, addr: Addr)

Sets the 'Source address' field

pub fn set_destination(&mut self, addr: Addr)

Sets the 'Destination address' field

pub fn payload_mut(&mut self) -> &mut [u8]

Immutable view into the payload

impl<B> Packet<B> where
B: AsMutSlice<Element = u8> + Truncate<u16>,

pub fn neighbor_advertisement(
&mut self,
target_ll_addr: Option<Addr>,
f: impl FnOnce(&mut Message<&mut [u8], NeighborAdvertisement>)
)

Fills the payload with a Neighbor Advertisement ICMPv6 message

pub fn udp(
&mut self,
f: impl FnOnce(&mut Packet<&mut [u8]>)
)

Fills the payload with a UDP packet

pub fn truncate(&mut self, len: u16)

Truncates the payload to the specified length