Struct jnet::sixlowpan::nhc::UdpPacket

pub struct UdpPacket<BUFFER> where
    BUFFER: AsSlice<Element = u8>,  { /* fields omitted */ }

LOWPAN_NHC compressed UDP packet

Methods

impl<B> UdpPacket<B> where
B: AsSlice<Element = u8>,

pub fn parse(buffer: B) -> Result<Self, B>

Parses the bytes as a LOWPAN_IPHC compressed IPv6 packet

pub fn get_source(&self) -> u16

Reads the (potentially compressed) 'Source Port' field

pub fn get_destination(&self) -> u16

Reads the (potentially compressed) 'Destination Port' field

pub fn get_checksum(&self) -> Option<u16>

Reads the (potentially compressed) 'Checksum' field

None means that the checksum has been elided by the compressor

pub fn payload(&self) -> &[u8]

Immutable view into the UDP payload

pub fn bytes(&self) -> &[u8]

Byte representation of this packet

pub fn get_c(&self) -> bool

Reads the 'Checksum' NHC field

pub fn get_p(&self) -> u8

Reads the 'Ports' NHC field

pub fn verify_ipv6_checksum(&self, src: Addr, dest: Addr) -> bool

Verifies the 'Checksum' field

impl<B> UdpPacket<B> where
B: AsMutSlice<Element = u8>,

pub fn payload_mut(&mut self) -> &mut [u8]

Mutable view into the payload

pub fn update_checksum(&mut self, src: Addr, dest: Addr)

Updates the checksum field, if not elided

impl<B> UdpPacket<B> where
B: AsMutSlice<Element = u8> + Truncate<u8>,

pub fn set_payload(&mut self, payload: &[u8])

Fills the payload with the given data and adjusts the length of the CoAP message

pub fn coap<F>(&mut self, token_length: u8, f: F) where
F: FnOnce(Message<&mut [u8], Unset>) -> Message<&mut [u8]>,

Fills the payload with a CoAP message