Struct jnet::udp::Packet

pub struct Packet<BUFFER> where
    BUFFER: AsRef<[u8]>,  { /* fields omitted */ }

UDP packet

Methods

impl<B> Packet<B> where
B: AsRef<[u8]>,

pub fn parse(bytes: B) -> Result<Self, B>

Parses the bytes as an UDP packet

pub fn get_source(&self) -> u16

Returns the Source (port) field of the header

pub fn get_destination(&self) -> u16

Returns the Destination (port) field of the header

pub fn get_length(&self) -> u16

Returns the Length field of the header

pub fn len(&self) -> u16

Returns the length (header + data) of this packet

pub fn payload(&self) -> &[u8]

View into the payload

pub fn as_bytes(&self) -> &[u8]

Returns the byte representation of this UDP packet

impl<B> Packet<B> where
B: AsRef<[u8]> + AsMut<[u8]>,

pub fn set_source(&mut self, port: u16)

Sets the Source (port) field of the header

pub fn set_destination(&mut self, port: u16)

Sets the Destination (port) field of the header

pub fn zero_checksum(&mut self)

Zeroes the Checksum field of the header

pub fn payload_mut(&mut self) -> &mut [u8]

Mutable view into the payload

impl<B> Packet<B> where
B: AsRef<[u8]> + AsMut<[u8]> + Resize,

pub fn new(buffer: B) -> Self

Transforms the given buffer into an UDP packet

NOTE The UDP packet will span the whole buffer and the Checksum field will be zeroed.

Panics

This constructor panics if the given buffer is not large enough to contain the UDP header.

pub fn set_payload(&mut self, data: &[u8])

Fills the payload with the given data and adjusts the length of the UDP packet

pub fn coap<F>(&mut self, token_length: u8, f: F) where
F: FnOnce(&mut Message<&mut [u8]>),

Fills the payload with a CoAP message

pub fn truncate(&mut self, len: u16)

Truncates the payload to the specified length