Struct jnet::arp::Packet

pub struct Packet<BUFFER, HTYPE = Ethernet, PTYPE = Ipv4> where
    BUFFER: AsSlice<Element = u8>,
    HTYPE: 'static,
    PTYPE: 'static,  { /* fields omitted */ }

ARP packet

Methods

impl<B> Packet<B, Ethernet, Ipv4> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8> + Truncate<u8>,

pub fn new(buffer: B) -> Self

Transforms the given buffer into an ARP packet

This function populates the following header fields:

HTYPE = Ethernet

PTYPE = IPv4

HLEN = 6

PLEN = 4

OPER = Request

impl<B> Packet<B, Ethernet, Ipv4> where
B: AsSlice<Element = u8>,

pub fn get_sha(&self) -> Addr

Returns the SHA (Sender Hardware Address) field of the payload

pub fn get_spa(&self) -> Addr

Returns the SPA (Sender Protocol Address) field of the payload

pub fn get_tha(&self) -> Addr

Returns the THA (Target Hardware Address) field of the payload

pub fn get_tpa(&self) -> Addr

Returns the TPA (Target Protocol Address) field of the payload

pub fn is_a_probe(&self) -> bool

Is this an ARP probe?

impl<B> Packet<B, Ethernet, Ipv4> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,

pub fn set_sha(&mut self, sha: Addr)

Sets the SHA (Sender Hardware Address) field of the payload

pub fn set_spa(&mut self, spa: Addr)

Sets the SPA (Sender Protocol Address) field of the payload

pub fn set_tha(&mut self, tha: Addr)

Sets the THA (Target Hardware Address) field of the payload

pub fn set_tpa(&mut self, tpa: Addr)

Sets the TPA (Target Protocol Address) field of the payload

pub fn announce(&mut self, addr: Addr)

ARP announcement

Shortcut for setting these fields

OPER = Request

SPA = TPA = addr

THA = 00:00:00:00:00:00

pub fn probe(&mut self, addr: Addr)

ARP probe

Shortcut for setting these fields

OPER = Request

SPA = 0.0.0.0

THA = 00:00:00:00:00:00

TPA = addr

impl<B> Packet<B, Unknown, Unknown> where
B: AsSlice<Element = u8>,

pub fn get_sha(&self) -> &[u8]

Returns the SHA (Sender Hardware Address) field of the payload

pub fn get_spa(&self) -> &[u8]

Returns the SPA (Sender Protocol Address) field of the payload

pub fn get_tha(&self) -> &[u8]

Returns the THA (Target Hardware Address) field of the payload

pub fn get_tpa(&self) -> &[u8]

Returns the TPA (Target Protocol Address) field of the payload

pub fn downcast(self) -> Result<Packet<B>, Self>

Interprets this packet as Packet<Ethernet, Ipv4>

impl<B> Packet<B, Unknown, Unknown> where
B: AsSlice<Element = u8>,

pub fn parse(bytes: B) -> Result<Self, B>

Parses bytes into an ARP packet

impl<B> Packet<B, Unknown, Unknown> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,

pub fn set_htype(&mut self, htype: HardwareType)

Sets the HTYPE (Hardware TYPE) field of the header

pub fn set_ptype(&mut self, ptype: Type)

Sets the PTYPE (Protocol TYPE) field of the header

impl<B, H, P> Packet<B, H, P> where
B: AsSlice<Element = u8>,

pub fn get_htype(&self) -> HardwareType

Returns the HTYPE (Hardware TYPE) field of the header

pub fn get_ptype(&self) -> Type

Returns the PTYPE (Protocol TYPE) field of the header

pub fn get_hlen(&self) -> u8

Returns the HLEN (Hardware LENgth) field of the header

pub fn get_plen(&self) -> u8

Returns the PLEN (Protocol LENgth) field of the header

pub fn get_oper(&self) -> Operation

Returns the OPER (OPERation) field of the header

pub fn payload(&self) -> &[u8]

View into the payload

NOTE this may contain padding bytes at the end

pub fn len(&self) -> u8

Returns the canonical length of this packet

This ignores padding bytes, if any

pub fn free(self) -> B

Frees the underlying buffer

impl<B, H, P> Packet<B, H, P> where
B: AsSlice<Element = u8> + AsMutSlice<Element = u8>,

pub fn set_oper(&mut self, oper: Operation)

Sets the OPER (OPERation) field of the header

pub fn payload_mut(&mut self) -> &mut [u8]

Mutable view into the payload

NOTE this may contain padding bytes at the end